2026 02 04 LW Fixed raw data recording (see desc)
Changed the mod_som_efe_obp_data_consumer_t field "record_timestamp" back into a 64-bit integer. It was accidentally changed to a 32-bit integer in a previous commit which led to "record_ptr" being overwritten whenever a new timestamp was saved. Also changed the mod_som_efe_obp_data_cpt_dissrate_t field "dissrates_cnt" into a 32-bit integer as 64 bits are not necessary for storing this value.

+++ mod/efe_obp/mod_som_efe_obp.h
@@ -190,7 +190,7 @@ typedef struct{
   uint32_t   spectrum_cnt;
   uint32_t   avgspec_cnt;
   uint32_t   rates_cnt;
-  uint32_t   record_timestamp;
+  uint64_t   record_timestamp;
 
   uint8_t *  record_ptr;     //ALB pointer to the segments section in the record
 
@@ -278,7 +278,7 @@ typedef struct{
 
   bool dof_flag;
   bool consumed_flag;
-  uint64_t dissrates_cnt;
+  uint32_t dissrates_cnt;
 //  uint64_t avg_timestamp[MOD_SOM_EFE_OBP_CPT_DISSRATE_NB_RATES_PER_RECORD];
   uint64_t avg_timestamp;
